refactor(transfer-state): move TransferStateProtocol into TransferState.swift and add interpolate method
